SDIODriver: Rename internal types/variables to reflect coming support for non-SDHCI controllers
A lot of state doesn't actually care whether a controller's register layout
follows the SDHCI standard or not. To reduce future confusion, most internal
types and pointer variables are hereby renamed from 'sdhci' to a more generic
'ctrl', short for 'controller'. Exceptions are sdhci_regset_t and
sdhci_writebuffer_t, which are intimately tied to register accesses.

Public data types and definitions (and the names of the exported header files)
retain the original names, for the sake of not breaking all the other
components that rely on them.