SDIODriver: Reorganise state variables internal to op_poll into separate struct
Three members of sdioop_t are particularly closely tied to the state machine
for SDHCI controllers, and so are likely to differ for soft controllers.
Create sdhci_op_t to hold these, and introduce a variable-length array member
at end of sdioop_t to hold it. Allocate sdioop_t more dyamically in
preparation for soft controllers.