SDIODriver: Refactor op_poll()
This is now split into an outer function that does locking, iterates over
the operation list and interacts with the OS and SDIODriver's clients, and
an inner function op_poll_sdhci() which codifies the wait states of a
standards-compliant SDHCI controller and does the register bashing. This
internal API is now pretty close to what is required for a soft controller.