SDIODriver: Add support for "soft" HAL devices
These are identified by a different value in the major version field, and are
intended for SD controllers that implement a register set that does not follow
the SDHCI standard. All interaction with the controller is performed via
function calls into such a HAL device, and a few of these are new.

A couple of other changes that were made in passing:

* device_added() now calls set_activity() and thereby avoids a branch through
  zero if a device indicates that the SDHCI host_control1 register should be
  used for setting the activity indicator.
* HAL device deactivate entry is now called before OS_ReleaseDeviceVector so
  that it can be used to mask interrupts in the host controller. Since no
  current versions of the module use interrupts, this is a safe change.