!Prepare: Fully remove Resources.UK.Intro file
This file was left over from the old version designed for CVS usage, where
components were held under a directory named according to the licence terms
used, and some of these components required copying to a different location
before they could be used. This is no longer the case, so this first step is
now obsolete.

The title text is moved into Resources.UK.Tools1, which is now the first file
to be printed.

Version 0.06. Tagged as 'PlingPrepare-0_06'