HdrSrc: Add a definition of the BEN bit in SCTLR (née MMU control register)
Detail:
The memory barriers ISB, DSB & DMB were intially introduced in ARMv6 as CP15
operations, however they were replaced at ARMv7 with dedicated instructions
and the old CP15 encodings deprecated. Since ARMv7VE was introduced, some CPUs
can be configured to make the CP15 encodings undefined instructions. For
unconfigurable CPUs, the CP15 encodings are always enabled, but for others,
we need to set this bit in SCTLR.

Note, this sort of behaviour usually implies ARM intend on removing these
instructions in a future architecture, so we might need to plan for this
eventuality.

Version 2.78. Tagged as 'HdrSrc-2_78'