HAL_Titanium: Fix stereo channels being reversed
By default RISC OS/SoundDMA generates 32bit sample frames where
the right channel is in the low 16 bits and the left channel is in
the high 16 bits, as required by VIDC's 'wide' sound FIFO which
stores a full 32bit frame per entry. This is in contrast to most
newer systems RISC OS supports, with 'narrow' FIFOs which store
only the data for a single channel per entry. Since I2S expects
the left channel to be transmitted first, this generally means
that RISC OS must switch to generating data such that the left
channel is in the low 16 bits of each word (as supported via the
audio HAL device 'StereoReverse' flag), or, as in this case, by
instructing the hardware to swap the channels for us.

Version 0.12. Tagged as 'HAL_Titanium-0_12'