HdrSrc: Add mixed-case versions of Hdr:Proc macros
Macros in this header file had inconsistent capitalisation. This is a legacy
of the earliest macros being `ENTRY` and `EXIT`. Macro names that are all
uppercase or all lowercase are a bad idea, because they can clash with opcodes
or assembler directives that are added at a later date: this is what happened
with `ENTRY` when objasm was developed, hence the introduction of the identical
`Entry` macro, and mass renaming across the source tree was required when it
switched to using objasm.

The other downside of all-uppercase or all-lowercase macro names is that it
makes it impossible to design a syntax colouring algorithm that is forward
compatible with new opcodes (and ARM continues to add new opcodes with almost
every new minor revision of the architecture).

Note that related macro names depend on the vintage of the header at which
they were introduced: compare `ALTENTRY` with `EntryS`.

Additional mixed-case versions of macros `ALTENTRY`, `EXIT`, `EXITS`, `EXITV`,
`EXITVC`, `EXITVS`, `FRAMLDR` and `FRAMSTR` are hereby introduced. The old
macro names are also retained (it would be a very large job to substitute them
throughout the source tree) but at least new code can now consistently use
mixed case.

The `FRAMCOM` and `FRAMSPL` macros are for internal use of the header file,
so uppercase versions of these did not need to be retained.

Also:
* Delete `ENTRY` macro: it's a long time since anything has been buildable
  using aasm, which is the only assembler that accepts this macro
* Correct cut-and-paste errors in the messages `ExitVC without EntryS` and
  `ExitVS without EntryS`

Version 2.83. Tagged as 'HdrSrc-2_83'