srcbuild: Support escaped spaces in Components file -options for cross-compilation
The cross-compiling version invokes `make` commands using `execve()`, so we
have the job of splitting the argument string into arguments (rather than an
intervening shell, as would be the case had we used `system()`). The previous
algorithm for splitting was too simplistic, simply searching for space
characters. Quoted space characters have become very common in Components
files in the years since this was last tested - frequently used for the
Kernel, for example - so this is quite important to support!

Version 0.56. Tagged as 'srcbuild-0_56'