Env: Make Env files work with default 640K again
Changes in Env 1.40-1.41 require GNU.gawk and amu within the same application
slot at the same time, which exceeds the default 640K WimpSlot. There isn't
actually any need for the awk script to run within a Makefile (it's not
testing the Makefile's behaviour directly) so by moving the lines up to
!Common, the memory usage is reduced.

Version 1.42. Tagged as 'Env-1_42'