Debugger: Fix -ve increment interactive mode *MemoryA
When stepping backwards in interactive mode the step (in R8) relied on modulo 32b arithmetic. As the (unsigned) addition always causes a carry out, with a 40b address this led the top 8 bits of the address to increment on every decrement of the bottom 32b.
Changed to sign extend to 64b before the add.

Version 2.06. Tagged as 'Debugger-2_06'