Wimp: Add Wimp_Extend 15, "read slot size"
Wimp_Extend 15 is equivalent to Wimp_SlotSize,-1,-1, with the exception
that the sizes returned in R0-R2 are now measured in pages instead of
bytes. This allows code to accurately read the amount of free memory on
machines with >2GB of RAM, and provides some future-proofing should we
ever wish to support tasks with >2GB of RAM (e.g. via a future
super-jumbo wimpslot, or more likely via sparse/PMP wimpslots, where
the physical size is larger than the logical size).

Currently there's no need to provide a page-based call to set the slot
size (~2GB is fine for the forseeable future), so this call only
focuses on reading the sizes.

Version 5.67. Tagged as 'Wimp-5_67'