HAL_OMAP3: Add missing BB rev C free to use table, correct the BBxM one
All versions of Beagleboard were referring to a single table, but at rev C three of the GPIO pins got moved, so we need a different table for that so that the GPIO module will enumerate the right one.
For Beagleboard xM, 10 pins were missing.
From discussion at
  https://www.riscosopen.org/forum/forums/11/topics/11903?page=2#posts-97755
Tested on Beagleboard C4 and xM.

Version 1.19. Tagged as 'HAL_OMAP3-1_19'