MemNow: Support over 2GB of RAM
Where supported, use Wimp_Extend 15 to read the amount of free memory
(in units of pages), in order to avoid hitting the 2GB limit imposed by
the byte-based Wimp_SlotSize.

Text buffer also enlarged, to avoid it overflowing once we have machines
with 10GB or more.

Version 0.08. Tagged as 'MemNow-0_08'