Switcher: Support over 2GB of RAM
This updates the Tasks window and related code so that memory amounts
are tracked in units of KB instead of bytes, allowing the code to
support machines with more than 2GB of RAM. Where relevant, newer
page-based OS APIs are used instead of older byte-based ones, which
should future-proof the code for operation on machines with over 4GB of
RAM (potentially up to the 1TB limit supported by the ARMv7 large
physical address extension).
Use definitions from Hdr:OSMem.

Version 1.52. Tagged as 'Switcher-1_52'