HAL_Titanium: Don't overfill FIFO above 38400 baud
The default FCR transmit FIFO threshold was being used (=8) but HAL_UARTFIFOSize was declaring 64 (the maximum supported). For baud rates > 19200bps DualSerial always tries to use the transmit FIFO (switch 'NewTXStrategy' in the sources) so was happily pouring 64 bytes in and losing 56.

UART.s: Enable enhanced functionality in EFR so that the FCR bits 4 & 5 can be written. Set them to 32, and declare that as the TX FIFO size to DualSerial.
RegMap: Add EFR bit definitions.

Version 0.13. Tagged as 'HAL_Titanium-0_13'