Wimp: Fix issue where dragging text from icon caused a crash in foreground task
When releasing claim, use of stale register meant that the foreground task wasn't correctly bouncing Message_Dragging when the pointer moved out of range. Also added a sanity check in get_selected_text so that if for some reason the selection window disappears mid action it will be caught and not cause a crash.

Version 5.71. Tagged as 'Wimp-5_71'