HAL_OMAP3: Fix UART TX FIFO thresholds
HAL_UARTFIFOSize was incorrectly returning the TX FIFO size, instead of
the number of characters that can be written when the TX empty IRQ is
received. This was causing data loss when DualSerial drives the
interface at high speed (> 19200 baud) (OMAP3530), or data aborts when
writing to UART_THR (DM3750, some kind of hardware quirk when the FIFO
is full?)

Fix this by programming a sensible trigger level at startup, and
returning the correct value from HAL_UARTFIFOSize.

Version 1.20. Tagged as 'HAL_OMAP3-1_20'