HAL_OMAP4: Fix UART TX FIFO thresholds
HAL_UARTFIFOSize was incorrectly returning the TX FIFO size, instead of
the number of characters that can be written when the TX empty IRQ is
received. This was causing data loss when DualSerial drives the
interface at high speed (> 19200 baud).

Fix this by programming a sensible trigger level at startup, and
returning the correct value from HAL_UARTFIFOSize.

Version 0.61. Tagged as 'HAL_OMAP4-0_61'