DWCDriver: Make a substitute for MPHI
The MPHI is (ab)used by DWCDriver as a means to do a FIQ downgrade to IRQ, but Pi 4 has no MPHI, so instead we substitute the GIC. This is essentially a GIC based equivalent to the changes in https://github.com/raspberrypi/linux/commit/0a1cf0d.
Apply these diffs then, in dwc_otg_riscos.c:
* deduce use_swirq by noticing the MPHI base happens to have all 13 LSB clear whereas the GIC doesn't, in just as horrendously Pi specific manner as Linux looking at the size of the memory region to deduce same
* setup the pointers to swirq_set and swirq_clr at GICD_ISPENDRn GICD_ICPENDRn

Version 0.36. Tagged as 'DWCDriver-0_36'