EtherUSB: Copy MAC in network byte order
EtherUSB stores the MAC internally in network byte order, except when the address is obtained via OS_ReadSysInfo 4, when it was stored reversed. This problem lay hidden because the only ROM in which EtherUSB is able to get an address that way is the Pi, whose HAL returned the address in the wrong order too.
Requires HAL_BCM2835-0_84.

Version 0.42. Tagged as 'EtherUSB-0_42'