PPP: Drop support for individual blockdriver/DeviceFS builds
PPP-1_45 introduced a bewildering 112 compile time switches to produce a 'COMBO' module which could select between using a blockdriver or DeviceFS. Since the resulting binary is almost the same size (55k v 53k) get rid of all the switches and just select at runtime.

Also fixed a couple of instances in passing where a loop iterated over NPPP but declared a pointer to ppp_softc[0] outside the loop, therefore not just repeatedly doing the action for the 0th interface.