PPP: Fix sensitivity to arg ordering
A bug introduced in PPP-1_45 meant that the order of command line arguments was significant because DriverType was always assigned in setdevname(). So when parse_args() speculatively tries each argument around line 300 a later speed or IP address would fail to parse, but also set the driver type back to BlockDriver.
Now, only assigned the driver type on success.

Version 1.46. Tagged as 'PPP-1_46'