EtherCPSW: Leave CPSW_SS_FLOW_CONTROL at its default value
Following the trail from where it appeared in NetBSD
  http://cvsweb.netbsd.org/bsdweb.cgi/src/sys/arch/arm/omap/if_cpsw.c.diff?r1=1.11&r2=1.12
mentions copying it from FreeBSD, where this changed added it
  https://svnweb.freebsd.org/base?view=revision&revision=246276
in an attempt to address a watchdog problem. Linux is the polar opposite and sets all the bits in the register (=7)
  https://github.com/torvalds/linux/blob/master/drivers/net/ethernet/ti/cpsw.c#L813
Given we never had a watchdog problem, and setting it to 0 seems to cause some egress problems
  https://www.riscosopen.org/forum/forums/11/topics/15154
play safe and put leave it with the SoC's default value (=1).