WinEdit: template_read_icon_name fix
Icons with no name can have a validation string of the form "N5//". This
passes all of the sanity checks for whether an icon name is present, but
then fails at the sscanf stage, with nothing being written to the output
buffer (not even a null terminator). If the output buffer hadn't been
zero-initialised by the caller then this would most likely result in
template_read_icon_name falsely claiming that a name is present.

Fixes garbled interactive help text in the "window properties" window
when hovering over the different labels ("Gadgets:", etc.)