WinEdit: Icon renumbering fix
When renumbering a list of icons, if the first few icons are already in
the correct position then:

1. The first icon will get swapped with itself, causing `where` to be
incremented from 0 to 1.
2. `i` will be set to `where`
3. `i` will get incremented by the `for` loop

`i` will now have the value 2, causing icon 1 to be skipped. This will
then cause icon 2 to get swapped with icon 1, and then icon 3 will get
swapped with icon 2 (the original icon 1), and then icon 4 will get
swapped with icon 3 (the original icon 1), and so on, shuffling icon 1
all the way to the end of the group, and needlessly disrupting the
numbering of all the other icons.

Fix by setting `i` to `where-1`, to take into account the increment that
the `for` loop is about to apply.

I'm not sure if the old logic would have resulted in broken/incomplete
sorting; mainly I'm aiming to fix the annoying behaviour where adding a
new icon to a window (which already had the icons correctly sorted)
would needlessly scramble the order of 90% of the icons.

The code's still a bit nasty (it should really shuffle the icons instead
of swap them), and it'll still want to shuffle everything if you add a
group box, but every little bit helps.