HAL_iMx6: Fix SMP doorbell device
With FIQs enabled, the doorbell interrupts were only being delivered as
FIQs, not IRQs. Fix so that they'll be delivered as both (the current
SMP module expects them to be IRQs, but FIQs could be a sensible future
improvement)

Fixes hang on Service_PagesUnsafe when SMP module is active (+
unreliable SMP thread scheduling), due to the doorbell interrupts never
triggering

Version 1.05. Tagged as 'HAL_iMx6-1_05'