Filer: [489] Fix for Ctrl-X/Ctrl-C clearing Filer rename writable
When Ctrl-X or Ctrl-C are pressed during a rename the Wimp (via Clipboard Manager) claim the clipboard to denote the text that was placed on it. This sends a Message_ClaimEntity. Unfortunately Filer wasn't discriminating which entity was being claimed, when it is only interested in loss of caret, so removed the writable and aborted the rename.
Add a check on the entity being claimed; only do something for caret.

Version 2.48. Tagged as 'Filer-2_48'