ABRelease: Tone down ZPP warning, document developer modules
First added in July 2015, no longer provide ZeroPain with every development ROM since it could be seen as a cause for alarm to a new user, and is a useful diagnosis tool to old deck hands.
Move the documentation of ZeroPain to its new home in the BonusBin, and add some notes and a link to the docs of the other developer modules provided (DebugBttn, DebugTools).

Version 1.00. Tagged as 'ABRelease-1_00'